// pronti-frontend-main/front/src/app/modulos/trabajador/form/form-beneficio.enum.ts
export enum FORM_BENEFICIO_ENUM {
  aporteIessAsumido = 'aporteIessAsumido',
  decimosAnio = 'decimosAnio',
  fondoReservaIess = 'fondoReservaIess',
  utilidades = 'utilidades',
  pagoFondosReservaMes = 'pagoFondosReservaMes',
  impuestoRentaPatron = 'impuestoRentaPatron',
  vacaciones = 'vacaciones',
  beneficios = 'beneficios',
  pagoDecTercerCuartoMes = 'pagoDecTercerCuartoMes'
}